pdf gears/requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
PyPDF2>=3.0.0
Pillow>=9.0.0
PyMuPDF>=1.20.0
python-docx>=0.8.0
pandas>=1.5.0
reportlab>=3.6.0
openpyxl>=3.0.0
pdf2image>=3.1.0
pytesseract>=0.3.10
pdf2docx>=0.5.0
camelot-py[cv]>=0.10.0
docx2pdf>=0.1.8
python-pptx>=0.6.21